Major refinment for lilypond ftplugin with the use of functions.
In the vimrc there is a new command that allows you to make simple
headers with a lot of ease. Also a few minor tweaks for lilypond.

Furthermore the Lilypond octave mappings (to add an octave above/below)
have finally been polished to the greatest extent. The regex is somewhat
daunting but now you don't need to worry about stuff inside <...> since
that will now be "ignored". So only notes that are outside of <...> will
now be affected by the octave mappings.